Associate locks with originating lock servers
This means it's no longer possible to release locks from servers
that don't know about the given lock.

Probably this could have been achieved using a functor, by relying
on the generative nature of functors to generate fresh locks and
server types, but the resulting interface is cleaner with the current
approach.